Refactor: Home화면 UI변경
- View시스템을 Compose로 마이그레이션
- UI 배치 및 컴포넌트 디자인 변경
