+++ core/designsystem/src/main/kotlin/com/hapataka/questwalk/core/designsystem/component/Button.kt
@@ -1,13 +1,23 @@
 package com.hapataka.questwalk.core.designsystem.component
 
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.interaction.collectIsPressedAsState
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.nativeCanvas
+import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.core.content.ContextCompat
@@ -42,6 +52,35 @@ fun PixelButton(
         Text(
             text = text,
             color = if (enabled) ButtonBrown else Color.White,
+        )
+    }
+}
+
+@Composable
+fun ImageButton(
+    releasedPainterResource: Painter,
+    contentDescription: String?,
+    pressedPainterResource: Painter? = null,
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit
+) {
+    val interactionSource = remember { MutableInteractionSource() }
+    val isPressed by interactionSource.collectIsPressedAsState()
+
+    Box(
+        modifier = modifier
+            .background(color = Color.Transparent)
+            .clickable(
+                onClick = onClick,
+                indication = null,
+                interactionSource = interactionSource,
+            ),
+    ) {
+        Image(
+            painter = if (isPressed) pressedPainterResource
+                ?: releasedPainterResource else releasedPainterResource,
+            contentDescription = contentDescription,
+            modifier = Modifier.fillMaxSize()
         )
     }
 }
